treehacksServer/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.6.1
blis==0.7.4
bs4==0.0.1
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
gensim==3.8.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
murmurhash==1.0.5
nltk==3.5
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pathlib==1.0.1
pathy==0.3.6
Pillow==8.1.0
preshed==3.0.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
smart-open==3.0.0
soupsieve==2.2
spacy==3.0.1
spacy-legacy==3.0.1
sqlparse==0.4.1
srsly==2.4.0
textblob==0.15.3
thinc==8.0.1
threadpoolctl==2.1.0
tokenizers==0.10.1
tqdm==4.56.2
transformers==4.3.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
zipp==3.4.0
